node count -> uint32

+++ include/node.h
@@ -3,11 +3,11 @@
 #include <stdint.h>
 
 struct mushroom_node {
-	uint16_t id;
+	uint32_t id;
 	char *address;
 };
 
-struct mushroom_node *mushroom_node_new(uint16_t id, const char *address);
+struct mushroom_node *mushroom_node_new(uint32_t id, const char *address);
 void mushroom_node_free(struct mushroom_node *node);
 
 struct mushroom_node *mushroom_node_copy(const struct mushroom_node *node);
